requirements.txt
click==8.1.3
python-box==7.1.1
pyyaml==5.4.1
rich==13.5.2
toposort==1.10
yarl==1.9.2

# plugin requirements - opotional
boto3==1.28.43
python-dotenv==0.21.1
# ngrok-api==0.10.0
virtualenv==20.24.5
